topic/fix for overflow

+++ leizd-aptos-logic/sources/risk_factor.move
@@ -299,7 +299,7 @@ module leizd_aptos_logic::risk_factor {
     }
     //// for round up
     fun calculate_fee_with_round_up(value: u64, fee: u64): u64 {
-        let value_mul_by_fee = (value * fee as u128); // NOTE: as not to overflow
+        let value_mul_by_fee = (value as u128) * (fee as u128); // NOTE: as not to overflow
         let precision_u128 = (precision() as u128);
         let result = value_mul_by_fee / precision_u128;
         if (value_mul_by_fee % precision_u128 != 0) (result + 1 as u64) else (result as u64)
